BMS\LTC6804.c - fixed some IO issues - reverted to SPI mode: MSB little endian - corrected endianess error in DATA and CMD/PEC
BMS\LTC6804_handler.h
- corrected type of Int_Temp to float

applications\app_maraneo.c
- added charge port current code (untested)
- added code for BMS state in bms_status command

+++ BMS/LTC6804_handler.h
@@ -48,7 +48,7 @@ typedef struct
 	
 	bool BMS_present;
 	
-	uint32_t Int_Temp;
+	float Int_Temp;
 
 	bool Cell_Test_Passed;
 	bool GPIO_Test_Passed;
